+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/bot/ktor/KtorCallFactory.kt
@@ -1,0 +1,15 @@
+package dev.inmo.tgbotapi.bot.ktor
+
+import dev.inmo.tgbotapi.requests.abstracts.Request
+import dev.inmo.tgbotapi.utils.TelegramAPIUrlsKeeper
+import io.ktor.client.HttpClient
+import kotlinx.serialization.json.Json
+
+interface KtorCallFactory {
+    suspend fun <T: Any> makeCall(
+        client: HttpClient,
+        urlsKeeper: TelegramAPIUrlsKeeper,
+        request: Request<T>,
+        jsonFormatter: Json
+    ): T?
+}